Changed spawn rate from int to float

+++ grizzly/context.py
@@ -163,7 +163,7 @@ class GrizzlyContextSetup:
     global_context: Dict[str, Any] = field(init=False, repr=False, hash=False, compare=False, default_factory=dict)
 
     user_count: int = field(init=False, default=0)
-    spawn_rate: Optional[int] = field(init=False, default=None)
+    spawn_rate: Optional[float] = field(init=False, default=None)
     timespan: Optional[str] = field(init=False, default=None)
 
     statistics_url: Optional[str] = field(init=False, default=None)
